uploaded spring task

+++ DLTE-Spring/Tasks/JDBCTemplate/src/main/java/jdbc/template/demo/controller/TransactionController.java
@@ -39,7 +39,7 @@ public class TransactionController {
     }
     @GetMapping("/amount/{amount}")
     //retrieving list of transactions by entering amount
-    public List<Transactions> findByAmount(@PathVariable Long amount) {
+    public List<Transactions> findByAmount(@PathVariable Double amount) {
         return transactionService.findByAmount(amount);
     }
 }
